Add operation to filter the cheapest carriers
The form to calculate the cheapest carrier is using km_price_cents. I've decided to continue with this simple metric because it is a big deal. That might be one of such important business rules. I have no data to decided and fell comfortable. I got many notes about the kinds of "freight tax". Like use the price per KG, sometimes it would be useful; share the trip cost for each customer (assess); Use the estimeted time to arrive to destination (we have urban via and road).
There are such ways to calculate it, as I've said. I just pick up one of them. But would be great implement/improve this rule, in the future.